Initial public commit
The actual initial commit was 2022-08-24. Given the large amount of code
in okitec's farmdec, I wondered whether it is possible to generate a
decoder/disassembler automatically from instruction descriptions.
Originally, I only implemented a trie-based classifier -- which is
extremely close to what is here now -- and a string-formatter, but no
operand decoder ("go and extract bits yourself") or encoder. The
classifier was extremely fast, but not very useful on its own.

A few months later, I was working on AArch64 code generation and was in
the need of a fast encoding library. Many years ago, I wrote a set of
macros to encode a few AArch64 instructions -- so why not auto-generate
this from the instruction descriptions? This was surprisingly easy,
although it required a different set of mnemonics: when decoding, "ADC"
is sufficient and not encoding operand sizes leads to fewer cases in
switch statements. For the encoding API, however, I wanted registers to
be plain integers and not carry (often useless) size information, so the
mnemonics for encoding are "ADCw" and "ADCx". Encoding of immediates is
non-trivial, especially for the logical operations AND/OR/EOR/ANDS/TST
(see da_immlogical in encode.c).

Another few months later (early 2024), we started working on the AArch64
back-end of our own compiler framework. This provided motivation to
actually completely cover the base ISA -- Advanced SIMD is a huge
instruction set -- and important extensions. It also resulted
performance improvements, most notably, the encoding functions are now
(by default) inline functions in the header. Thus, the encoder has an
extremely low performance cost.

Now I had a more-or-less complete instruction encoding description,
classifier, string-formatter (which was intensively fuzz-tested against
LLVM-MC), and encoder ... but no usable decoder. So I went and annotated
every instruction class with decode operand information and replaced the
string-based formatter with an operand-based formatter. I also realized,
how many operand types AArch64 has, especially in comparison to, e.g.,
x86, which has around 10. The classifer can classify an instruction in
about ~20 instructions and the operand decoder executes another 50-60
instructions on average (x86-64 and AArch64, libLLVM.so) and can execute
with an IPC of >2 (on both, Apple M1 and an older x86-64 Core). I think
this is reasonably fast and it is acceptable to say that decoding is
"fast enough". The formatter still has some room for perf improvements.

Why did I decide to scrap 1.8 years of commit history? It had some
lengthy holes and bulk commits, which are mostly an artefact of being a
side project for some time (also reflects in the brevity of most commit
messages). I also don't think there's valuable information in that
history, and code that was removed was replaced with something that was
superior in functionality. So here we are. Thank you for reading. <3